adding tests for the create model endpoint (#39)
* adding tests for the create model endpoint

and fixing it. the tests involve adding an interface for the list
creator

* use validateAndCrate for the model resource

* Allowing for plural model names in the new endpoint

* Adding logging to the model resource

* Adding ID cols for all the join tables

If there’s no ‘id’ column, fizz automatically assigns one as an int. In
many cases pop will try to put a uuid (from one of the models in the
join table) into that column, causing an error

* Using a lower-case name for gifm

With all upper-case, pop put underscores in between each character,
causing lots of issues with table names

* Redirecting to the singular model name

* Fixing column name for gifm <> author join table

* updates to the new and create tests

* Final fix up to the model resource test